+++ flask_via/__init__.py
@@ -58,17 +58,17 @@ class RoutesImporter(object):
 
 
 class Via(RoutesImporter):
-    """ The core class which kicks off the whole registration processes.
+    """ Flask-VIa integration into Flask applications. Flask-Via can
+    be integrated in two different ways depending on how you have setup your
+    Flask application.
 
     .. versionadded:: 2014.05.06
 
-    Example
-    -------
-    .. sourcecode:: python
+    You can bind to a specific flask application::
 
         from flask import Flask
         from flask.ext.via import Via
-        from flask.ext.via.routers.flask import Basic
+        from flask.ext.via.routers.flask import Functional
 
         app = Flask(__name__)
 
@@ -76,16 +76,57 @@ class Via(RoutesImporter):
             return 'Foo View!'
 
         routes = [
-            Basic('/foo', foo),
-            Basic('/foo/<bar>', foo, endpoint='foo2'),
+            Functional('/foo', foo),
+            Functional('/foo/<bar>', foo, endpoint='foo2'),
         ]
 
+        via = Via(app, routes_module='path.to.here')
+
+        if __name__ == "__main__":
+            app.run(debug=True)
+
+    Or if you use an application factory you can use
+    :meth:`flask_via.Via.init_app`::
+
+        from flask import Flask
+        from flask.ext.via import Via
+        from flask.ext.via.routers.flask import Functional
+
         via = Via()
-        via.init_app(app, routes_module='path.to.here')
+
+        def foo(bar=None):
+            return 'Foo View!'
+
+        routes = [
+            Functional('/foo', foo),
+            Functional('/foo/<bar>', foo, endpoint='foo2'),
+        ]
+
+        def create_app():
+            app = Flask(__name__)
+            via.init_app(app)
+            return app
+
+        app = create_app()
 
         if __name__ == "__main__":
             app.run(debug=True)
     """
+
+    def __init__(self, app=None, *args, **kwargs):
+        """ Constructor. Functionalally acts as a proxy to
+        :meth:`flask_store.Store.init_app`.
+
+        .. versionadded:: 2014.05.19.2
+
+        Key Arguments
+        -------------
+        app : flask.app.Flask, optional
+            Optional Flask application instance, default None
+        """
+
+        if app:
+            self.init_app(app, *args, **kwargs)
 
     def init_app(
             self,
@@ -104,6 +145,10 @@ class Via(RoutesImporter):
               app configuration variable. If ``routes_name`` keyword argument
               and ``VIA_ROUTES_NAME`` are not configured the default will be
               routes.
+
+        .. versionchanged:: 2014.05.19.2
+
+            * Improved ``init_app`` method
 
         Arguments
         ---------
@@ -128,17 +173,16 @@ class Via(RoutesImporter):
             and ``route_module`` keyword argument has not been provided.
         """
 
-        if not routes_module:
-            routes_module = app.config.get('VIA_ROUTES_MODULE')
+        app.config.setdefault('VIA_ROUTES_MODULE', routes_module)
+        app.config.setdefault('VIA_ROUTES_NAME', routes_name or 'routes')
 
-        if not routes_module:
+        if not app.config['VIA_ROUTES_MODULE']:
             raise ImproperlyConfigured(
                 'VIA_ROUTES_MODULE is not defined in application '
                 'configuration.')
 
-        # Routes name can be configured by setting VIA_ROUTES_NAME
-        if not routes_name:
-            routes_name = app.config.get('VIA_ROUTES_NAME', 'routes')
+        routes_module = app.config['VIA_ROUTES_MODULE']
+        routes_name = app.config['VIA_ROUTES_NAME']
 
         # Get the routes
         routes = self.include(routes_module, routes_name)