+++ CHANGELOG.rst
@@ -1,6 +1,12 @@
 Change Log
 ==========
 
+2014.05.19.2
+------------
+* Improved: @joonathan added ``kwargs`` passing to the ``Blueprint`` Router
+* Improved: ``init_app`` and added ``__init__``
+* Added: Deprecation warning to ``Basic`` router
+
 2014.05.19.1
 ------------
 * Hotfix: Fixed issue where routes would be reregistered with an app